try to connect to database

+++ backend/db/create_tables.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS TRIAL (
     id INT AUTO_INCREMENT PRIMARY KEY,
     selection_type VARCHAR(50),
-    duration INT,
-    distance_from_center INT,
+    duration BIGINT,
+    distance_from_center DOUBLE,
     circle_radius FLOAT,
-    space_between INT
+    space_between FLOAT
 )
